+++ src/Algebra/Ceil.php
@@ -42,15 +42,15 @@ final class Ceil implements Number
     }
 
     #[\Override]
-    public function add(Number $number, Number ...$numbers): Number
+    public function add(Number $number): Number
     {
-        return Addition::of($this, $number, ...$numbers);
+        return Addition::of($this, $number);
     }
 
     #[\Override]
-    public function subtract(Number $number, Number ...$numbers): Number
+    public function subtract(Number $number): Number
     {
-        return Subtraction::of($this, $number, ...$numbers);
+        return Subtraction::of($this, $number);
     }
 
     #[\Override]
@@ -60,9 +60,9 @@ final class Ceil implements Number
     }
 
     #[\Override]
-    public function multiplyBy(Number $number, Number ...$numbers): Number
+    public function multiplyBy(Number $number): Number
     {
-        return Multiplication::of($this, $number, ...$numbers);
+        return Multiplication::of($this, $number);
     }
 
     #[\Override]