+++ src/Algebra/CommonLogarithm.php
@@ -47,15 +47,15 @@ final class CommonLogarithm implements Number
     }
 
     #[\Override]
-    public function add(Number $number, Number ...$numbers): Number
+    public function add(Number $number): Number
     {
-        return Addition::of($this, $number, ...$numbers);
+        return Addition::of($this, $number);
     }
 
     #[\Override]
-    public function subtract(Number $number, Number ...$numbers): Number
+    public function subtract(Number $number): Number
     {
-        return Subtraction::of($this, $number, ...$numbers);
+        return Subtraction::of($this, $number);
     }
 
     #[\Override]
@@ -65,9 +65,9 @@ final class CommonLogarithm implements Number
     }
 
     #[\Override]
-    public function multiplyBy(Number $number, Number ...$numbers): Number
+    public function multiplyBy(Number $number): Number
     {
-        return Multiplication::of($this, $number, ...$numbers);
+        return Multiplication::of($this, $number);
     }
 
     #[\Override]