+++ src/Algebra/Factorial.php
@@ -48,15 +48,15 @@ final class Factorial implements Number
     }
 
     #[\Override]
-    public function add(Number $number, Number ...$numbers): Number
+    public function add(Number $number): Number
     {
-        return $this->collapse()->add($number, ...$numbers);
+        return $this->collapse()->add($number);
     }
 
     #[\Override]
-    public function subtract(Number $number, Number ...$numbers): Number
+    public function subtract(Number $number): Number
     {
-        return $this->collapse()->subtract($number, ...$numbers);
+        return $this->collapse()->subtract($number);
     }
 
     #[\Override]
@@ -66,9 +66,9 @@ final class Factorial implements Number
     }
 
     #[\Override]
-    public function multiplyBy(Number $number, Number ...$numbers): Number
+    public function multiplyBy(Number $number): Number
     {
-        return $this->collapse()->multiplyBy($number, ...$numbers);
+        return $this->collapse()->multiplyBy($number);
     }
 
     #[\Override]